requirements.txt
# Primary packages.
bayesian-optimization==1.2.0
numpy==1.20.3
pyyaml==6.0
scikit-learn==0.24.2
scipy==1.7.1
tensorflow==2.8.0

# Sub-packages that the primary packages use.
absl-py==1.3.0
astunparse==1.6.3
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
flatbuffers==22.12.6
gast==0.5.3
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
idna==3.4
importlib-metadata==5.1.0
joblib==1.2.0
keras==2.8.0
keras-preprocessing==1.1.2
libclang==14.0.6
markdown==3.4.1
markupsafe==2.1.1
oauthlib==3.2.2
opt-einsum==3.3.0
protobuf==4.21.11
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-io-gcs-filesystem==0.28.0
termcolor==2.1.1
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
typing-extensions==4.4.0
urllib3==1.26.13
werkzeug==2.2.2
wheel==0.38.4
wrapt==1.14.1
zipp==3.11.0